Added examples to test blurring an image
hello-world.example tests GaussianBlurView
Button 1 deactivates, button 2 removes the view from stage, and button 3 adds it back and renders for 2 seconds

image-view-blur.example tests RenderTasks directly.
There are no controls for this demo.

First, it creates a scene, setting the image to load immediately, and adds it to the stage, but sets
it's visibility to false so it doesn't render.

It waits for the image to finish loading, then it creates a chain of render tasks as follows:
The scene is rendered onto FB 1,
FB 1 is rendered into a preview FB 2 in the top left.
FB 1 is rendered into FB 3 using the gaussian blur shader horizontally.
FB 3 is rendered into a preview FB 4 in the bottom left
FB 3 is rendered into FB 1 using the gaussian blur shader vertically,
FB 1 is rendered into the final image view in the center.

Pressing 1 removes the render tasks, previews and camera actors, and grows the final image.

Change-Id: I0167657e87ee13342c57a61d984d09c2cbff6b93
Signed-off-by: David Steele <david.steele@samsung.com>