Added examples to test blurring an image
hello-world.example tests GaussianBlurView
Button 1 deactivates, Button 2 Activates,
Button 3 destroys, Button 4 recreates.

image-view-blur.example tests RenderTasks directly.
It uses both a Gaussian blur and downsampling (to mimic GaussianBlurView)

First, it creates a scene, setting the image to load immediately, and adds it to the stage, but sets
it's visibility to false so it doesn't render.

It waits for the image to finish loading, then it creates a chain of render tasks as follows:
The scene is rendered onto FB 1,
FB 1 is rendered into a preview FB 2 in the top left.
FB 1 is rendered into FB 3 using the gaussian blur shader horizontally, FB 3 is half-size
FB 3 is rendered into a preview FB 4 in the bottom left
FB 3 is rendered into FB 4 using the gaussian blur shader vertically, FB 4 is same size as FB 3
FB 4 is upsampled into FB 1
FB 1 is rendered into the final image view in the center.

Pressing 1 removes the render tasks, previews and camera actors, and grows the final image.

Change-Id: I0167657e87ee13342c57a61d984d09c2cbff6b93
Signed-off-by: David Steele <david.steele@samsung.com>