Added Colour Transition example.
Demonstrates colour transition using flow maps and uv rotation / scaling.

Flow maps are greyscale images where the value of the pixels signifies the
progress of the animation, which is added to an animated value which we
use to lerp between old and new colour.

The colour of the content is used to scale the source / target colour, i.e.
white is affected most, dark is affected less.

Controls:
- Double-tap middle (33%) of screen: transition using random flow map;
- Double-tap outside the middle of screen: transition using one of the
  effects (flow maps) mapped to the given segment of the screen;

Change-Id: I8767d7632f30e489f9546c0f862bcdd05bb0305a
Signed-off-by: György Straub <g.straub@partner.samsung.com>