Add window example
<Summary>
DALi is a rendering engine of NUI and was previously a public interface that can be used by applications, but now it is only used as an internal engine, so most verifications are verified by NUI sample and UTC. This makes it difficult to verify whether it is DALi's own issue or NUI issue.

<What>
A window sample was added to the DALi-demo so that the APIs provided by DALi can be verified. This patch contains the DALi Window Function function.

<Why>
Other modules of the existing DALi had samples in the dali-demo, so it was possible to verify the function in the DALi layer, but the window did not have the sample, so it was difficult to verify the DALi window itself. In addition, the Window module itself had a strong dependence on external modules such as Display Server, AppFW, and Sensor FW, so it had no choice but to perform limited verification to the existing UTC.
To solve this problem, I am adding the sample.

<Plan>
Now the patch is only provided with samples for Window Function. We also plan to add function verification for Event or Signal.

Change-Id: I415a8795684964a492f1a229f17e129a02903724