[Tizen] Added 2d physics benchmark
First 30s, runs N balls using Animation & property notification that don't
interact,
Second 30s, runs N balls using 2D Physics and collisions.

N by default is 500, but can be changed on the command line.
FPS tracking is automatically switched on for this benchmark

(Also changed CMakeLists.txt to be able to work as a subdirectory
in a multi-repo build)

Change-Id: Iccee1d40650563e3f45e021b4fb3fff871cc393e