Optimized benchmark animation (Reduce event thread overhead)
Since we were create new animations every PropertyNotify, and at that time,
we iterate whole balls.

If event thread time delayed (~= the ball is very much) event callback
stacked a lot, and then break the app.

To avoid it, we make the number of animations relative with ball +
change the animation only if we need, not always.

It will reduce the event thread overhead.

Change-Id: Ic6da267e827d012c9e68c7d64594596f3617233a
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>